Feature/authorization

+++ Banana_King/Pages/BananaOperations.cshtml.cs
@@ -1,12 +1,16 @@
 using Banana_King.Models;
 using Banana_King.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace Banana_King.Pages
 {
+    [Authorize]
     public class BananaOperationsModel : PageModel
     {
+        public bool IsAdmin => HttpContext.User.HasClaim("IsAdmin", bool.TrueString);
+
         public List<Banana>? Bananas = new();
 
         [BindProperty]
@@ -19,6 +23,8 @@ namespace Banana_King.Pages
 
         public IActionResult OnPost()
         {
+            if (!IsAdmin) return Forbid();
+
             if (!ModelState.IsValid)
             {
                 return Page();
@@ -30,6 +36,7 @@ namespace Banana_King.Pages
 
         public IActionResult OnPostDelete(int id)
         {
+            if (!IsAdmin) return Forbid();
             BananaService.Delete(id);
             return RedirectToAction("Get");
         }
